feat!: add compatibility for forward-auth middleware
BREAKING CHANGE: The forward-auth middleware is expected to extract
authentication from `request`, say, based on headers, and inject a valid
Strapi user authentication into the request prior to processing of other
middlewares. Such a middleware is, however, not part of this commit.

The admin dashboard is crippled in the sense that authentication is
assumed to always be present (and empty!), since the responsibility of
authentication is shifted into the back-end middleware. The log-out
button is adjusted to redirect to `/oauth2/sign_out`.

This commit leaves the admin dashboard non-functinal without a
forward-auth middleware present in the final Strapi application.

(cherry picked from commit 60ca12a6ed7ca1e54d8ff051d2d6788c1fccf9b4)