Remove token checks
Setting the frontend state token is not possible from the forward-auth middleware, therefore the frontend will be permanently unlocked to recreate the previous patch versions.
This entails always returning website content and requesting user data even if no token is set in the frontend, as it will be injected via middleware later.