Patch increasing ID upon update published
Whenever an already published entry is updated, the draft gets republished. This essentialy means the old published entry is deleted and a new one is created, resulting in a changed ID for the published version of a document.
This change updates the old published entry instead of deleting it.
The dist versions of the affected file have been provided for easier patch creation.