Handle separate publishing
Publish is called everytime a published entry is updated, but can also be used to simply publish a draft, in which case no data is provided to the function (as it would normally just copy the draft and update its status to published).
Instead, a data object is extracted from the draft to update the published entry with.